requirement.txt
bleach==1.5.0
certifi==2016.2.28
click==6.7
cycler==0.10.0
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
numpy==1.14.3
opencv-python==3.4.1.15
pandas==0.20.3
Pillow==5.1.0
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
scipy==1.1.0
six==1.11.0
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.8
tornado==4.5.2
visitor==0.1.3
Werkzeug==0.14.1
